+++ requirements.txt
@@ -1,17 +1,10 @@
-dacite==1.8.1
-Flask==3.0.3
-fschat==0.2.36
-ftfy==6.2.0
-openai==1.37.1
-pandas==2.2.2
-pyserini==0.35.0
-python-dotenv==1.0.1
-PyYAML==6.0.1
-PyYAML==6.0.1
-Requests==2.32.3
-setuptools==69.5.1
-tiktoken==0.7.0
-torch==2.3.1
-tqdm==4.66.4
-transformers==4.42.4
-vllm==0.5.2
+tqdm>=4.66.2
+openai>=1.23.6
+tiktoken>=0.6.0
+transformers>=4.40.1
+pyserini>=0.35.0
+python-dotenv>=1.0.1
+faiss-cpu>=1.8.0
+ftfy>=6.2.0
+dacite>=1.8.1
+fschat[model_worker]>=0.2.36
