+++ src/rank_llm/__init__.py
@@ -1,3 +1,3 @@
-from .retrieve_and_rerank import extract_kwargs
+from .retrieve_and_rerank import extract_kwargs, retrieve_and_rerank
 
-__all__ = ["extract_kwargs"]
+__all__ = ["extract_kwargs", "retrieve_and_rerank"]
