+++ src/rank_llm/api/server.py
@@ -4,16 +4,9 @@ import torch
 from flask import Flask, jsonify, request
 
 from rank_llm import retrieve_and_rerank
-from rank_llm.rerank.listwise.gpt.api_keys import (
-    get_azure_openai_args,
-    get_openai_api_key,
-)
-from rank_llm.rerank.listwise.listwise_rankllm import PromptMode
-from rank_llm.rerank.rank_gpt import SafeOpenai
-from rank_llm.rerank.rank_listwise_os_llm import RankListwiseOSLLM
-from rank_llm.rerank.reranker import OperationMode
-from rank_llm.retrieve.pyserini_retriever import RetrievalMethod
-from rank_llm.retrieve.retriever import RetrievalMode
+from rank_llm.rerank import get_azure_openai_args, get_openai_api_key, PromptMode
+from rank_llm.rerank.listwise import SafeOpenai, RankListwiseOSLLM
+from rank_llm.retrieve import RetrievalMethod, RetrievalMode
 
 """ API URL FORMAT
 
@@ -117,7 +110,7 @@ def create_app(model, port, use_azure_openai=False):
             default_agent = None
         try:
             # calls Anserini retriever API and reranks
-            (response, agent) = retrieve_and_rerank.retrieve_and_rerank(
+            (response, agent) = retrieve_and_rerank(
                 dataset=dataset,
                 retrieval_mode=RetrievalMode.DATASET,
                 query=query,
@@ -132,7 +125,6 @@ def create_app(model, port, use_azure_openai=False):
                 num_passes=num_passes,
                 retrieval_method=_retrieval_method,
                 print_prompts_responses=False,
-                operation_mode=OperationMode.STANDARD.value,
             )
 
             # set the default reranking agent to the most recently used reranking agent
