+++ src/rank_llm/demo/experimental_results.py
@@ -8,17 +8,13 @@ sys.path.append(parent)
 
 from rank_llm.analysis.response_analysis import ResponseAnalyzer
 from rank_llm.evaluation.trec_eval import EvalFunction
-from rank_llm.rerank.listwise.gpt.api_keys import get_openai_api_key
-from rank_llm.rerank.rank_gpt import SafeOpenai
-from rank_llm.rerank.reranker import Reranker
-from rank_llm.rerank.vicuna_reranker import VicunaReranker
-from rank_llm.rerank.zephyr_reranker import ZephyrReranker
+from rank_llm.rerank.listwise import SafeOpenai, VicunaReranker, ZephyrReranker
+from rank_llm.rerank import get_openai_api_key
 from rank_llm.retrieve.retriever import Retriever
 from rank_llm.retrieve.topics_dict import TOPICS
 
 # create rerankers
-agent = SafeOpenai("gpt-3.5-turbo", 4096, keys=get_openai_api_key())
-g_reranker = Reranker(agent)
+g_reranker = SafeOpenai("gpt-3.5-turbo", 4096, keys=get_openai_api_key())
 v_reranker = VicunaReranker()
 z_reranker = ZephyrReranker()
 rerankers = {"rg": g_reranker, "rv": v_reranker, "rz": z_reranker}
