+++ src/rank_llm/demo/rerank_dataset_with_custom_index.py
@@ -6,8 +6,8 @@ parent = os.path.dirname(SCRIPT_DIR)
 parent = os.path.dirname(parent)
 sys.path.append(parent)
 
-from rank_llm.rerank.vicuna_reranker import VicunaReranker
-from rank_llm.retrieve.retriever import Retriever
+from rank_llm.rerank.listwise import VicunaReranker
+from rank_llm.retrieve import Retriever
 
 # Data prep: download the pyserini prebuilt index and topics files
 os.system(
