+++ src/rank_llm/rerank/__init__.py
@@ -1,10 +1,11 @@
 import logging
 
 from .identity_reranker import IdentityReranker
-from .rankllm import RankLLM
+from .rankllm import RankLLM, PromptMode
+from .api_keys import get_azure_openai_args, get_openai_api_key
 
 logging.basicConfig(
     level=logging.INFO, format="%(asctime)s - %(levelname)s - %(message)s"
 )
 
-__all__ = ["IdentityReranker", "RankLLM"]
+__all__ = ["IdentityReranker", "RankLLM", "get_azure_openai_args", "get_openai_api_key", "PromptMode"]
