+++ src/rank_llm/rerank/listwise/gpt/__init__.py
@@ -1,9 +1,6 @@
-from .api_keys import get_azure_openai_args, get_openai_api_key
 from .rank_gpt import CompletionMode, SafeOpenai
 
 __all__ = [
     "SafeOpenai",
     "CompletionMode",
-    "get_azure_openai_args",
-    "get_openai_api_key",
 ]
