+++ src/rank_llm/rerank/listwise/gpt/rank_gpt.py
@@ -7,7 +7,8 @@ import tiktoken
 from tqdm import tqdm
 
 from rank_llm.data import Request, Result
-from rank_llm.rerank.listwise import ListwiseRankLLM, PromptMode
+from rank_llm.rerank import PromptMode
+from rank_llm.rerank.listwise import ListwiseRankLLM
 
 
 class CompletionMode(Enum):
@@ -184,31 +185,17 @@ class SafeOpenai(ListwiseRankLLM):
         return [
             {
                 "role": "system",
-                "content": "You are RankGPT, an advanced assistant specialized in ranking passages by their relevance to a given query.",
+                "content": "You are RankGPT, an intelligent assistant that can rank passages based on their relevancy to the query.",
             },
             {
                 "role": "user",
-                "content": f"You will be given {num} passages, marked with a numerical identifier []. Rank these passages according to how relevant they are to the query: {query}.",
+                "content": f"I will provide you with {num} passages, each indicated by number identifier []. \nRank the passages based on their relevance to query: {query}.",
             },
             {"role": "assistant", "content": "Okay, please provide the passages."},
         ]
 
-        # return [
-        #     {
-        #         "role": "system",
-        #         "content": "You are RankGPT, an intelligent assistant that can rank passages based on their relevancy to the query.",
-        #     },
-        #     {
-        #         "role": "user",
-        #         "content": f"I will provide you with {num} passages, each indicated by number identifier []. \nRank the passages based on their relevance to query: {query}.",
-        #     },
-        #     {"role": "assistant", "content": "Okay, please provide the passages."},
-        # ]
-
     def _get_suffix_for_rank_gpt_prompt(self, query: str, num: int) -> str:
-        # APEER
-        return f"Search Query: {query}.\nArrange the {num} passages above in order of relevance to the search query, from most to least relevant. Use the numerical identifiers for ranking. The format should be [] > [], e.g., [1] > [2]. Only provide the ranking results without any additional text or explanation."
-        # return f"Search Query: {query}. \nRank the {num} passages above based on their relevance to the search query. The passages should be listed in descending order using identifiers. The most relevant passages should be listed first. The output format should be [] > [], e.g., [1] > [2]. Only response the ranking results, do not say any word or explain."
+        return f"Search Query: {query}. \nRank the {num} passages above based on their relevance to the search query. The passages should be listed in descending order using identifiers. The most relevant passages should be listed first. The output format should be [] > [], e.g., [1] > [2]. Only response the ranking results, do not say any word or explain."
 
     def _get_prefix_for_rank_gpt_apeer_prompt(
         self, query: str, num: int
