+++ src/rank_llm/rerank/listwise/vicuna_reranker.py
@@ -1,7 +1,8 @@
 from typing import List
 
 from rank_llm.data import Request, Result
-from rank_llm.rerank.listwise import PromptMode, RankListwiseOSLLM
+from rank_llm.rerank import PromptMode
+from rank_llm.rerank.listwise import RankListwiseOSLLM
 
 
 class VicunaReranker:
