+++ src/rank_llm/rerank/rankllm.py
@@ -1,12 +1,20 @@
 import logging
 from abc import ABC, abstractmethod
-from typing import Any, Dict, List, Tuple, Union
+from typing import Any, Dict, List, Enum, Tuple, Union
 
 from rank_llm.data import Request, Result
 
 logger = logging.getLogger(__name__)
 
+class PromptMode(Enum):
+    UNSPECIFIED = "unspecified"
+    RANK_GPT = "rank_GPT"
+    RANK_GPT_APEER = "rank_GPT_APEER"
+    LRL = "LRL"
 
+    def __str__(self):
+        return self.value
+    
 class RankLLM(ABC):
     def __init__(
         self,
