+++ src/rank_llm/retrieve_and_rerank.py
@@ -1,13 +1,10 @@
 from typing import Any, Dict, List, Union
 
 from rank_llm.data import Query, Request
-from rank_llm.rerank import IdentityReranker, RankLLM
+from rank_llm.rerank import IdentityReranker, RankLLM, PromptMode, get_azure_openai_args, get_openai_api_key
 from rank_llm.rerank.listwise import (
-    PromptMode,
     RankListwiseOSLLM,
     SafeOpenai,
-    get_azure_openai_args,
-    get_openai_api_key,
 )
 from rank_llm.retrieve import (
     RetrievalMethod,
